Lots o changes.
Added in sentry logging.

Fixed the timestamp flicker.

Move the chat room to it's own browser tab instead of nested into tabs in one window.
Big win here as the layout of the template is much cleaner, there there is still
some janitorial work to be done.

User list now updates as the application polls account for people leaving and coming.

Added in a title bar that account for Company Name and Chat channel name.

Pulled out motd to its own seperate div so that it is always in one spot.

Bolded all names, though this may be changed in the future.

chatbox container is now a well div so that it looks different from the other pieces
of the application.