Various changes and bug fixes.
changed the connected users div to a panel, looks nicer.

fixed the scroll bar stuck at the top bug.  upon joining it now
auto scrolls to the bottom, once the initial load is done the user
has full control over scroll, and it will only autoscroll it it is at
the bottom (i.e. when a new message comes is).

put in titleAlert to alert of new messages in a channel.

BACKEND:

Began converting everything over to a Javascript framework currently being written.

This can be seen in the ajax calls slowly disappearing from channel.js/chat.js
and being replaced with calls to an object api.  This makes the functional js code
much cleaner and easier to read while also cleaning up the ajax calls and defining
them all in one place as opposed to all over.  first on the list to conver is
anything that touches or deals with channels.